+++ app/src/main/java/com/github/se/gomeet/ui/mainscreens/create/Create.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.ButtonDefaults
@@ -19,6 +20,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontStyle
@@ -42,8 +44,17 @@ import com.github.se.gomeet.ui.theme.DarkCyan
  */
 @Composable
 fun Create(nav: NavigationActions) {
-  Scaffold(
+    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
+    val screenHeight = LocalConfiguration.current.screenHeightDp.dp
+
+    Scaffold(
       modifier = Modifier.testTag("CreateUI"),
+      topBar = {
+          Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.padding(start = screenWidth/15, top = screenHeight/30)) {
+              Text(text = "Create",
+                  style = MaterialTheme.typography.headlineMedium.copy(fontWeight = FontWeight.SemiBold))
+              Spacer(Modifier.weight(1f))
+          } },
       bottomBar = {
         BottomNavigationMenu(
             onTabSelect = { selectedTab ->
@@ -52,73 +63,55 @@ fun Create(nav: NavigationActions) {
             tabList = TOP_LEVEL_DESTINATIONS,
             selectedItem = Route.CREATE)
       }) { innerPadding ->
-        Text(
-            text = "Create",
-            modifier = Modifier.padding(horizontal = 15.dp, vertical = 15.dp),
-            color = DarkCyan,
-            fontStyle = FontStyle.Normal,
-            fontWeight = FontWeight.SemiBold,
-            fontFamily = FontFamily.Default,
-            textAlign = TextAlign.Start,
-            style = MaterialTheme.typography.headlineLarge)
-
-        Column(
+      Column(
             modifier = Modifier.fillMaxSize().padding(innerPadding),
             verticalArrangement = Arrangement.Top,
             horizontalAlignment = Alignment.CenterHorizontally) {
-              Spacer(modifier = Modifier.height(150.dp))
+              Spacer(modifier = Modifier.height(screenHeight/6))
 
               Text(
                   text = "Choose your audience",
-                  fontStyle = FontStyle.Normal,
-                  fontWeight = FontWeight.Normal,
-                  fontFamily = FontFamily.Default,
-                  textAlign = TextAlign.Start,
-                  style = MaterialTheme.typography.headlineSmall)
+                  style = MaterialTheme.typography.headlineSmall
+              )
 
-              Spacer(modifier = Modifier.height(150.dp))
-
-              Row {
-                OutlinedButton(
-                    modifier = Modifier.width(128.dp).height(40.dp).testTag("CreatePublic"),
-                    shape = RoundedCornerShape(10.dp),
-                    border = BorderStroke(1.dp, Color.Gray),
-                    enabled = true,
-                    colors = ButtonDefaults.outlinedButtonColors(containerColor = Color.LightGray),
-                    onClick = {
-                      nav.navigateTo(CREATE_ITEMS.first { it.route == Route.PUBLIC_CREATE })
-                    }) {
+          Spacer(modifier = Modifier.height(screenHeight/6))
+                  OutlinedButton(
+                      onClick = {
+                          nav.navigateTo(
+                              CREATE_ITEMS.first { it.route == Route.PUBLIC_CREATE })
+                      },
+                      modifier = Modifier
+                          .width((screenWidth / 1.5.dp).dp)
+                          .height(screenHeight / 17),
+                      shape = RoundedCornerShape(10.dp),
+                      enabled = true,
+                      colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.outlineVariant)){
                       Text(
                           text = "Public",
-                          color = Color.Black,
-                          fontStyle = FontStyle.Normal,
-                          fontWeight = FontWeight.Normal,
-                          fontFamily = FontFamily.Default,
-                          textAlign = TextAlign.Start,
-                          style = MaterialTheme.typography.bodyMedium)
-                    }
-                Spacer(modifier = Modifier.width(20.dp))
+                          style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.SemiBold),
+                          color = Color.White)
+                  }
 
-                OutlinedButton(
-                    modifier = Modifier.width(128.dp).height(40.dp),
-                    shape = RoundedCornerShape(10.dp),
-                    border = BorderStroke(1.dp, Color.Gray), // Set border here if needed
-                    enabled = true,
-                    colors = ButtonDefaults.outlinedButtonColors(containerColor = Color.LightGray),
-                    onClick = {
-                      nav.navigateTo(CREATE_ITEMS.first { it.route == Route.PRIVATE_CREATE })
-                    }) {
+                    Spacer(modifier = Modifier.size(10.dp))
+                  OutlinedButton(
+                      onClick = {
+                          nav.navigateTo(
+                              CREATE_ITEMS.first { it.route == Route.PRIVATE_CREATE })
+
+                      },
+                      modifier = Modifier
+                          .width((screenWidth / 1.5.dp).dp)
+                          .height(screenHeight / 17),
+                      shape = RoundedCornerShape(10.dp),
+                      border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant),
+                      enabled = true,
+                      colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.background)) {
                       Text(
                           text = "Private",
-                          color = Color.Black,
-                          fontStyle = FontStyle.Normal,
-                          fontWeight = FontWeight.Normal,
-                          fontFamily = FontFamily.Default,
-                          textAlign = TextAlign.Start,
-                          style = MaterialTheme.typography.bodyMedium)
-                    }
+                          style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.SemiBold),
+                          color = MaterialTheme.colorScheme.outlineVariant)
+                  }
               }
-            }
       }
 }
